# swagger.yaml
openapi: 3.0.3
info:
    title: Coffee Supply Management System API
    version: "1.0.0"
    contact:
        name: Team 67 - PZSP2
        email: coffee-system@pw.edu.pl
...
paths:
    /auth/login:
        post:
            tags: [Authentication]
            summary: Log in as coordinator
            operationId: login
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/LoginRequest"
                        example:
                            username: admin
                            password: secret
            responses:
                "200":
                    description: Session token
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/LoginResponse"
                "401":
                    $ref: "#/components/responses/Unauthorized"
    /auth/register:
        post:
            tags: [Authentication]
            summary: Register a new coordinator
            operationId: register
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/LoginRequest"
            responses:
                "201":
                    description: Account created
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/LoginResponse"
                "409":
                    description: Username already taken
...
    /distributors/{distributor_id}/api-keys:
        get:
            tags: [API Keys]
            summary: List distributor API keys
            operationId: listApiKeys
            security:
                - BearerAuth: []
            parameters:
                - $ref: "#/components/parameters/DistributorId"
            responses:
                "200":
                    description: List of API keys
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: "#/components/schemas/ApiKeyResponse"

        post:
            tags: [API Keys]
            summary: Generate API key for distributor
            operationId: createApiKey
            security:
                - BearerAuth: []
            parameters:
                - $ref: "#/components/parameters/DistributorId"
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/ApiKeyCreateRequest"
                        example:
                            label: "ERP Integration Key"
            responses:
                "201":
                    description: Key generated
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ApiKeyResponse"
                            example:
                                id: "3f2b1a..."
                                key: "cof_xK7mNpQ..."
                                label: "ERP Integration Key"
                                distributor_id: "dist-001"
                                active: true
                                created_at: "2026-04-16T10:00:00Z"
    /api-keys/{key_id}:
        delete:
            tags: [API Keys]
            summary: Revoke API key
            operationId: revokeApiKey
            security:
                - BearerAuth: []
            parameters:
                - name: key_id
                  in: path
                  required: true
                  schema:
                      type: string
                      format: UUID
            responses:
                "204":
                    description: Key revoked
                "401":
                    $ref: "#/components/responses/Unauthorized"
                "404":
                    $ref: "#/components/responses/NotFound"
    /buildings:
        get:
            tags: [Buildings]
            summary: List buildings
            operationId: listBuildings
            security:
                - BearerAuth: []
            responses:
                "200":
                    description: List of buildings
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: "#/components/schemas/BuildingResponse"

        post:
            tags: [Buildings]
            summary: Add building
            operationId: createBuilding
            security:
                - BearerAuth: []
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/BuildingCreateRequest"
                        example:
                            name: "Building A - Downtown"
                            location: "123 Main St, Warsaw"
                            max_capacity_kg: 150.0
                            initial_inventory_kg: 20.0
                            daily_demand:
                                - day: 1
                                  demand_kg: 15.0
                                - day: 2
                                  demand_kg: 18.0
            responses:
                "201":
                    description: Building created
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/BuildingResponse"
    /buildings/{building_id}:
        get:
            tags: [Buildings]
            summary: Building details
            operationId: getBuilding
            security:
                - BearerAuth: []
            parameters:
                - $ref: "#/components/parameters/BuildingId"
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/BuildingResponse"
                "404":
                    $ref: "#/components/responses/NotFound"

        put:
            tags: [Buildings]
            summary: Update building
            operationId: updateBuilding
            security:
                - BearerAuth: []
            parameters:
                - $ref: "#/components/parameters/BuildingId"
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/BuildingCreateRequest"
            responses:
                "200":
                    description: Updated
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/BuildingResponse"

        delete:
            tags: [Buildings]
            summary: Delete building
            operationId: deleteBuilding
            security:
                - BearerAuth: []
            parameters:
                - $ref: "#/components/parameters/BuildingId"
            responses:
                "204":
                    description: Deleted
...
components:
    securitySchemes:
        BearerAuth:
            type: http
            scheme: bearer
            bearerFormat: JWT
            description: Coordinator session token (from `/auth/login`)
        ApiKeyAuth:
            type: apiKey
            in: header
            name: X-Api-Key
            description: Distributor API key (format `cof_...`)
    parameters:
        DistributorId:
            name: distributor_id
            in: path
            required: true
            schema:
                type: string
                format: UUID
            description: Distributor UUID
        BuildingId:
            name: building_id
            in: path
            required: true
            schema:
                type: string
                format: UUID
            description: Building UUID
    responses:
        Unauthorized:
            description: Missing or invalid token/API key
            content:
                application/json:
                    schema:
                        $ref: "#/components/schemas/ErrorResponse"
                    example:
                        detail: "Authentication required"
        NotFound:
            description: Resource not found
            content:
                application/json:
                    schema:
                        $ref: "#/components/schemas/ErrorResponse"
                    example:
                        detail: "Not found"
    schemas:
        ErrorResponse:
            type: object
            properties:
                detail:
                    type: string
        LoginRequest:
            type: object
            required: [username, password]
            properties:
                username:
                    type: string
                password:
                    type: string
                    format: password
        LoginResponse:
            type: object
            properties:
                token:
                    type: string
                    description: Bearer token for use in Authorization header
                user_id:
                    type: string
                    format: UUID
                role:
                    type: string
                    enum: [coordinator, admin]
...
        DailyDemand:
            type: object
            required: [day, demand_kg]
            properties:
                day:
                    type: integer
                    minimum: 1
                demand_kg:
                    type: number
                    format: float
                    minimum: 0
                    description: "Daily demand [kg]"
        BuildingCreateRequest:
            type: object
            required: [name, max_capacity_kg, daily_demand]
            properties:
                name:
                    type: string
                location:
                    type: string
                    nullable: true
                max_capacity_kg:
                    type: number
                    format: float
                    description: "Storage capacity [kg]"
                initial_inventory_kg:
                    type: number
                    format: float
                    default: 0.0
                    description: "Initial inventory level [kg]"
                daily_demand:
                    type: array
                    items:
                        $ref: "#/components/schemas/DailyDemand"
        BuildingResponse:
            allOf:
                - $ref: "#/components/schemas/BuildingCreateRequest"
                - type: object
                  properties:
                      id:
                          type: string
                      current_inventory_kg:
                          type: number
                          format: float
                          description: "Current inventory level [kg]"
                      created_at:
                          type: string
                          format: date-time
                      updated_at:
                          type: string
                          format: date-time
...
        ApiKeyCreateRequest:
            type: object
            required: [label]
            properties:
                label:
                    type: string
                    description: Descriptive key label
        ApiKeyResponse:
            type: object
            properties:
                id:
                    type: string
                key:
                    type: string
                    description: "API key (format: cof_...) — only present on creation"
                label:
                    type: string
                distributor_id:
                    type: string
                    format: UUID
                active:
                    type: boolean
                revoked_at:
                    type: string
                    format: date-time
                    nullable: true
                created_at:
                    type: string
                    format: date-time